Split three Key instances modules out of Cardano.Api.Typed
The Byron, Shelley and Praos key types and Key class instances.
